rag/embeddings.py
# rag/embeddings.py
from sentence_transformers import SentenceTransformer
from typing import List
import numpy as np

class Embedder:
    def __init__(self, model_name: str = "all-MiniLM-L6-v2"):
        self.model = SentenceTransformer(model_name)

    def embed(self, texts: List[str]) -> np.ndarray:
        """
        Returns a float32 numpy array of shape (len(texts), dim)
        """
        embeddings = self.model.encode(texts, convert_to_numpy=True, show_progress_bar=False)
        # ensure float32 for faiss compatibility
        return np.array(embeddings, dtype="float32")
