requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
numpy
requests
groq
